+++ src/constants/status.rs
@@ -48,10 +48,10 @@ impl Display for Flags {
 #[derive(Debug)]
 pub struct Status {
     pub level: u8,
-    pub strength: u16,
-    pub agility: u16,
-    pub max_hp: u16,
-    pub max_mp: u16,
+    pub strength: u8,
+    pub agility: u8,
+    pub max_hp: u8,
+    pub max_mp: u8,
     pub required_exp: u16,
     pub spell: Option<&'static str>,
 }
@@ -60,8 +60,8 @@ pub struct Status {
 pub struct PlayerSummary {
     pub name: String,
     pub level: u8,
-    pub hp: u16,
-    pub mp: u16,
+    pub hp: u8,
+    pub mp: u8,
     pub gold: u16,
     pub experience: u16,
 }
@@ -69,12 +69,12 @@ pub struct PlayerSummary {
 #[derive(Debug)]
 pub struct StrengthStatus {
     pub level: u8,
-    pub strength: u16,
-    pub agility: u16,
-    pub max_hp: u16,
-    pub max_mp: u16,
-    pub attack_power: u16,
-    pub defense_power: u16,
+    pub strength: u8,
+    pub agility: u8,
+    pub max_hp: u8,
+    pub max_mp: u8,
+    pub attack_power: u8,
+    pub defense_power: u8,
     pub weapon: String,
     pub armor: String,
     pub shield: String,
